% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contLikMLE2.R
\name{contLikMLE2}
\alias{contLikMLE2}
\title{contLikMLE2}
\usage{
contLikMLE2(
  nC,
  samples,
  popFreq,
  refData = NULL,
  condOrder = NULL,
  knownRef = NULL,
  kit = NULL,
  AT = 50,
  pC = 0.05,
  lambda = 0.01,
  fst = 0,
  mixProp = 1:length(samples),
  PHexp = 1:length(samples),
  PHvar = 1:length(samples),
  DEG = rep(0, length(samples)),
  stuttBW = rep(0, length(samples)),
  stuttFW = rep(0, length(samples)),
  minF = NULL,
  normalize = FALSE,
  difftol = 0.01,
  steptol = 1e-06,
  nDone = 3,
  maxThreads = 0,
  delta = 1,
  seed = NULL,
  verbose = TRUE,
  knownRel = NULL,
  ibd = NULL
)
}
\arguments{
\item{nC}{Number of contributors in model}

\item{samples}{A List with samples which for each samples has locus-list elements with list elements adata and hdata. 'adata' is a qualitative (allele) data vector and 'hdata' is a quantitative (peak heights) data vector.}

\item{popFreq}{A list of allele frequencies for a given population.}

\item{refData}{Reference objects with (2-size) allele-vector given in list element [[i]][[s]].}

\item{condOrder}{Specify conditioning references from refData (must be consistent order). For instance condOrder=(0,2,1,0) means that we restrict the model such that Ref2 and Ref3 are respectively conditioned as 2. contributor and 1. contributor in the model.}

\item{knownRef}{Specify known non-contributing references from refData (index). For instance knownRef=(1,2) means that reference 1 and 2 is known non-contributor in the hypothesis. This affectes coancestry correction.}

\item{kit}{A vector of kitname for each sample name, used to model degradation. Must be one of the shortnames of kit: check getKit()}

\item{AT}{Analytical threshold. A vector per sample, or a list with marker elements per sample}

\item{pC}{Dropin prob parameter. A vector per sample, or a list with marker elements per sample}

\item{lambda}{DropinPH parameter. A vector per sample, or a list with marker elements per sample}

\item{fst}{The co-ancestry coefficient. Default is 0. A vector per marker.}

\item{mixProp}{Mixture proportion params: Indicate which shared parameter group each samples belongs to}

\item{PHexp}{PeakHeight expectation params: Indicate which shared parameter group each samples belongs to}

\item{PHvar}{PeakHeight variability params: Indicate which shared parameter group each samples belongs to}

\item{DEG}{Degradation slope params: Indicate which shared parameter group each samples belongs to}

\item{stuttBW}{Backward stutter params: Indicate which shared parameter group each samples belongs to}

\item{stuttFW}{Forward stutter params: Indicate which shared parameter group each samples belongs to}

\item{minF}{The freq value included for new alleles (new alleles as potential stutters will have 0). Default NULL is using min.observed in popFreq.}

\item{normalize}{A boolean of whether normalization should be applied or not. Default is FALSE.}

\item{difftol}{Tolerance for being exact in log-likelihood value (relevant when nDone>1)}

\item{steptol}{Argument used in the nlm function for faster return from the optimization (tradeoff is lower accuracy).}

\item{nDone}{Number of optimizations required providing equivalent results (same logLik value obtained)}

\item{maxThreads}{Maximum number of threads to be executed by the parallelization}

\item{delta}{Scaling of variation of normal distribution when drawing random startpoints. Default is 1.}

\item{seed}{The user can set seed if wanted}

\item{verbose}{Boolean whether printing optimization progress. Default is TRUE.}

\item{knownRel}{gives the index of the reference which the 1st unknown is related to.}

\item{ibd}{the identical by decent coefficients of the relationship (specifies the type of relationship)}
}
\value{
ret A list from model fit
}
\description{
contLikMLE optimizes the likelihood function of the DNA mixture model
}
\author{
Oyvind Bleka
}
