% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efm2.R
\name{efm2}
\alias{efm2}
\title{efm2}
\usage{
efm2(envirfile = NULL, envir = NULL)
}
\arguments{
\item{envirfile}{A Rdata file including a saved environment of a project}

\item{envir}{A environment object}
}
\description{
efm2 is a GUI wrapper for the EFMrep package
}
\author{
Oyvind Bleka
}
