% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPCAclusters.R
\name{getPCAclusters}
\alias{getPCAclusters}
\title{getPCAclusters}
\usage{
getPCAclusters(X, grps = NULL, txt = "", getClusters = TRUE, alpha = 0.1)
}
\arguments{
\item{X}{summed PH per marker per sample}

\item{grps}{Indicate which group each sample belongs to}

\item{txt}{Title name}

\item{getClusters}{Whether clusters should be estimated}

\item{alpha}{Coverage grade for ellipsoid of each cluster}
}
\value{
ret Cluster classification of each sample
}
\description{
Perform PCA on samples
}
\author{
Oyvind Bleka
}
