% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReportText.R
\name{getReportText}
\alias{getReportText}
\title{getReportText}
\usage{
getReportText(set, envir, sig = 2, colps = "\\t")
}
\arguments{
\item{set}{fitted MLE object (includes hd and possibly also hp)}

\item{envir}{environment object}

\item{sig}{significant numbers present for numeric}

\item{colps}{column separator to use}
}
\value{
txt return text of report
}
\description{
Helpfunction to get report text in efm2
}
\author{
Oyvind Bleka
}
