% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatednessSelectorGUI.R
\name{kinshipSelectorGUI}
\alias{kinshipSelectorGUI}
\title{kinshipSelectorGUI}
\usage{
kinshipSelectorGUI(refs, Crange, hypsel = "hd", env = NULL)
}
\arguments{
\item{refs}{Name of references}

\item{Crange}{range of contributors to allow}

\item{hypsel}{name of hypothesis to consider (obtain object from env)}

\item{env}{an environment object from efm2}
}
\description{
GUI window to specify kinship
}
\author{
Oyvind Bleka
}
