% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLiki2.R
\name{logLiki2}
\alias{logLiki2}
\title{logLiki2}
\usage{
logLiki2(mlefit)
}
\arguments{
\item{mlefit}{Fitted object using contLikMLE}
}
\value{
ret A vector with log-likelihood-values for each locus for given model
}
\description{
logLiki returns the likelihood for each markers (based on the MLE results)
}
\author{
Oyvind Bleka
}
