% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareC2.R
\name{prepareC2}
\alias{prepareC2}
\title{prepareC2}
\usage{
prepareC2(dat, repNames, nC, condOrder = NULL, knownRef = NULL,
  kit = NULL, AT = 50, pC = 0.05, lambda = 0.01, fst = 0,
  incBS = FALSE, incFS = FALSE, knownRel = NULL, ibd = NULL)
}
\arguments{
\item{dat}{A list which contains the output from function prepareData2 (samples,refs,freq)}

\item{repNames}{Names of replicate samples}

\item{nC}{Number of contributors in model (must be a consitant, same for all replicates)}

\item{condOrder}{Specify conditioning references from refData (must be consistent order). For instance condOrder=(0,2,1,0) means that we restrict the model such that Ref2 and Ref3 are respectively conditioned as 2. contributor and 1. contributor in the model.}

\item{knownRef}{Specify known non-contributing references from refData (indices). For instance knownRef=(1,2) means that reference 1 and 2 is known non-contributor in the hypothesis. This affectes coancestry correction.}

\item{kit}{shortname of kit: Obtained from getKit().}

\item{AT}{Analytical threshold. A vector per sample, or a list with marker elements per sample}

\item{pC}{Dropin prob parameter. A vector per sample, or a list with marker elements per sample}

\item{lambda}{DropinPH parameter. A vector per sample, or a list with marker elements per sample}

\item{fst}{The co-ancestry coefficient. Default is 0. A vector per marker.}

\item{incBS}{A boolean whether potential BW stutters are included}

\item{incFS}{A boolean whether potential FW stutters are included}

\item{knownRel}{Specify the index of the related contributing reference from refData (one index). For instance knownRel=2 means that unknown1 is related to reference 2 with ibd specified relationship.}

\item{ibd}{the identical by decent coefficients list of the relationship for each unknowns (specifies the type of relationship). Default is NULL, meaning no related inds. Can be a list for each defined related pairs.}
}
\value{
ret A list of data input to call the C-code with
}
\description{
prepareC is used in the functions contLikMLE to prepare input to C-call
}
\details{
Assumes that all PH below threshold are removed. The function builds the data input to the C-code
}
\author{
Oyvind Bleka
}
